Migrated protected tree methods to util class
