+++ src/index.ts
@@ -170,3 +170,4 @@ export { EntitySchemaEmbeddedColumnOptions } from "./entity-schema/EntitySchemaE
 export { ColumnType } from "./driver/types/ColumnTypes"
 export { EntitySchemaOptions } from "./entity-schema/EntitySchemaOptions"
 export { InstanceChecker } from "./util/InstanceChecker"
+export { TreeRepositoryUtils } from "./util/TreeRepositoryUtils"
